+++ src/components/Navbar.tsx
@@ -20,7 +20,6 @@ const Navbar = () => {
     { title: 'Impression 3D', path: '/3d-printing' },
     { title: 'IA', path: '/ai-experts' },
     { title: 'YouTubeurs Tech', path: '/tech-youtubers' },
-    { title: 'Microsoft Office', path: '/resources/microsoft-office' },
     { title: 'Resources', path: '/resources' },
   ];